Fixed WndProc
-Fixed a bug in WndProc, improved it a little.
-Made keyboard hook and mouse hook disabled by default.
-Made ProcessSharp use a dispose pattern

+++ src/Process.NET/Windows/WndProcEventArgs.cs
@@ -4,7 +4,7 @@ namespace Process.NET.Windows
 {
     public class WndProcEventArgs : EventArgs
     {
-        public WndProcEventArgs(IntPtr hwnd, int msg, int wParam, IntPtr lParam)
+        public WndProcEventArgs(IntPtr hwnd, int msg, IntPtr wParam, IntPtr lParam)
         {
             Hwnd = hwnd;
             Msg = msg;
@@ -16,7 +16,7 @@ namespace Process.NET.Windows
 
         public int Msg { get; }
 
-        public int WParam { get; }
+        public IntPtr WParam { get; }
 
         public IntPtr LParam { get; }
     }
